split components in files
