+++ src/main/js/screens/app/screens/quizes/QuizesScreen.js
@@ -1,10 +1,9 @@
 import React from 'react';
-import {
-  Alert, Grid, Row, Col, Form, FormGroup, FormControl, Button,
-  ButtonGroup, Panel, ButtonToolbar
-} from 'react-bootstrap';
+import { Alert, Grid, FormControl, Button, ButtonGroup } from 'react-bootstrap';
 
 import quizService from '../../../../shared/services/quizService';
+import QuizBox from './QuizBox';
+import QuizForm from './QuizForm';
 
 class QuizesScreen extends React.Component {
   constructor(props) {
@@ -97,7 +96,7 @@ class QuizesScreen extends React.Component {
         {
           this.state.showCreate
             ?
-            <CreateQuiz callParent={this.handleCreateQuiz}/>
+            <QuizForm callParent={this.handleCreateQuiz}/>
             :
             <p>
               <Button bsStyle="success" onClick={() => this.setState({ showCreate: true })}>Skapa en ny</Button><br/>
@@ -142,178 +141,4 @@ class QuizesScreen extends React.Component {
 }
 
 export default QuizesScreen;
-
-class QuizBox extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      showButtons: false,
-      quizDeleted: false,
-      deleteAlertVisible: true
-    };
-
-    this.onInputChange = this.onInputChange.bind(this);
-    this.mouseEnter = this.mouseEnter.bind(this);
-  }
-
-  onInputChange(event) {
-    this.setState({
-      [event.target.name]: event.target.value,
-    });
-  }
-
-  mouseEnter() {
-    this.setState({ showButtons: true });
-  }
-
-  mouseExit = () => {
-    this.setState({ showButtons: false });
-  }
-
-  handleDeleteQuiz(id) {
-    this.props.handleDeleteQuiz(id);
-  }
-
-  deleteQuiz(id) {
-    quizService().delete(id).then((response) => {
-      if (response.status === 200) {
-        this.setState({
-          quizDeleted: true
-        });
-      } else {
-        throw new Error();
-      }
-    }).catch((err) => {
-      this.setState({
-        error: 'Något fick fel'
-      });
-    });
-  }
-
-  render() {
-    let quiz = this.props.quiz;
-    return (
-      <div>
-        {this.state.quizDeleted ? (
-          <Alert bsStyle="info" onDismiss={() => this.handleDeleteQuiz(quiz.id) }>
-            <strong>{quiz.name}</strong> har raderats.
-          </Alert>
-        ) : (
-          <Panel
-            onMouseEnter={this.mouseEnter}
-            onMouseLeave={this.mouseExit}>
-            <h3>{quiz.name}</h3>
-            <p>{quiz.description}</p>
-            {this.state.showButtons ?
-              <ButtonToolbar>
-                <Button bsStyle="primary">Visa</Button>
-                <Button bsStyle="danger" onClick={() => this.deleteQuiz(quiz.id)}>Ta bort</Button>
-              </ButtonToolbar>
-              :
-              null
-              }
-          </Panel>
-        )}
-      </div>
-    );
-  }
-}
-
-class CreateQuiz extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      name: '',
-      description: '',
-      error: null,
-    };
-
-    this.onInputChange = this.onInputChange.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-  }
-
-  handleSubmit(event) {
-    event.preventDefault();
-    const obj = { name: this.state.name, description: this.state.description };
-    quizService().create(obj).then((response) => {
-      if (response.status === 201) {
-        this.props.callParent(true);
-      } else {
-        throw new Error();
-      }
-    }).catch((err) => {
-      this.setState({
-        error: 'Något fick fel',
-      });
-    });
-  }
-
-  renderErrorMsg() {
-    if (this.state.error != null) {
-      return (
-        <Alert bsStyle="danger">
-          <strong>Ett fel upptäcktes!</strong>
-          <p>{this.state.error}</p>
-        </Alert>
-      );
-    }
-    return '';
-  }
-
-  onInputChange(event) {
-    this.setState({
-      [event.target.name]: event.target.value,
-    });
-  }
-
-  render() {
-    return (
-      <Panel>
-        <h3>Skapa en ny quiz</h3>
-        <Form onSubmit={this.handleSubmit}>
-          <FormGroup>
-            <FormControl
-              type="text"
-              name="name"
-              placeholder="Ange ett quiznamn"
-              value={this.state.name}
-              onChange={this.onInputChange}
-            />
-          </FormGroup>
-          <FormGroup>
-            <FormControl
-              type="text"
-              name="description"
-              placeholder="Ange en beskrivning"
-              value={this.state.description}
-              onChange={this.onInputChange}
-            />
-          </FormGroup>
-          <FormGroup>
-            {this.renderErrorMsg()}
-            <ButtonGroup vertical block>
-              <Button
-                type="submit"
-                label="create"
-                bsStyle="primary"
-                name="create"
-                disabled={!this.state.name || !this.state.description}
-              >
-                Skapa
-              </Button>
-              <Button
-                label="cancel"
-                bsStyle="default"
-                name="cancel"
-                onClick={() => this.props.callParent(false)}
-              >
-                Ångra
-              </Button>
-            </ButtonGroup>
-          </FormGroup>
-        </Form>
-      </Panel>
-    );
-  }
-}
